// src/generated/openapi/models/ErrorResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * File Manager API
 * File Manager API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Status } from './Status';
import {
    StatusFromJSON,
    StatusFromJSONTyped,
    StatusToJSON,
} from './Status';
import type { ErrorType } from './ErrorType';
import {
    ErrorTypeFromJSON,
    ErrorTypeFromJSONTyped,
    ErrorTypeToJSON,
} from './ErrorType';

/**
 * 
 * @export
 * @interface ErrorResponse
 */
export interface ErrorResponse {
    /**
     * 
     * @type {Status}
     * @memberof ErrorResponse
     */
    status: Status;
    /**
     * 
     * @type {string}
     * @memberof ErrorResponse
     */
    description: string;
    /**
     * 
     * @type {ErrorType}
     * @memberof ErrorResponse
     */
    error: ErrorType;
}



/**
 * Check if a given object implements the ErrorResponse interface.
 */
export function instanceOfErrorResponse(value: object): value is ErrorResponse {
    if (!('status' in value) || value['status'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('error' in value) || value['error'] === undefined) return false;
    return true;
}

export function ErrorResponseFromJSON(json: any): ErrorResponse {
    return ErrorResponseFromJSONTyped(json, false);
}

export function ErrorResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ErrorResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'status': StatusFromJSON(json['status']),
        'description': json['description'],
        'error': ErrorTypeFromJSON(json['error']),
    };
}

export function ErrorResponseToJSON(value?: ErrorResponse | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'status': StatusToJSON(value['status']),
        'description': value['description'],
        'error': ErrorTypeToJSON(value['error']),
    };
}