// src/generated/openapi/models/ErrorType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * File Manager API
 * File Manager API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ErrorType = {
    BadRequest: 'bad_request',
    NotFound: 'not_found',
    ServerError: 'server_error'
} as const;
export type ErrorType = typeof ErrorType[keyof typeof ErrorType];


export function instanceOfErrorType(value: any): boolean {
    for (const key in ErrorType) {
        if (Object.prototype.hasOwnProperty.call(ErrorType, key)) {
            if (ErrorType[key as keyof typeof ErrorType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ErrorTypeFromJSON(json: any): ErrorType {
    return ErrorTypeFromJSONTyped(json, false);
}

export function ErrorTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ErrorType {
    return json as ErrorType;
}

export function ErrorTypeToJSON(value?: ErrorType | null): any {
    return value as any;
}